Generate assets hashes manually for internals usage
* sha1 hahses
* Fixes #129
