+++ lib/index.js
@@ -277,9 +277,10 @@ var OfflinePlugin = (function () {
         // By some reason errors raised here are not fatal,
         // so we need manually try..catch and exit with error
         try {
-          _this2.hash = compilation.hash;
           _this2.setAssets(compilation);
           _this2.setHashesMap(compilation);
+          // Probably generate bundle hash manually (from what we have)
+          _this2.hash = _loaderUtils2['default'].getHashDigest(Object.keys(_this2.hashesMap).join(''), 'sha1');
 
           // Not used yet
           // this.setNetworkOptions();
@@ -433,12 +434,6 @@ var OfflinePlugin = (function () {
     value: function setHashesMap(compilation) {
       var _this4 = this;
 
-      var hashesMap = this.findAssetsHashes(compilation);
-      var hashedAssets = Object.keys(hashesMap).reduce(function (result, hash) {
-        result[hashesMap[hash]] = hash;
-        return result;
-      }, {});
-
       this.hashesMap = {};
 
       Object.keys(compilation.assets).forEach(function (key) {
@@ -446,29 +441,10 @@ var OfflinePlugin = (function () {
 
         if (typeof validatedPath !== 'string' || _this4.assets.indexOf(validatedPath) === -1) return;
 
-        var hash = undefined;
-
-        if (hashedAssets[key]) {
-          hash = hashedAssets[key];
-        } else {
-          hash = _loaderUtils2['default'].getHashDigest(compilation.assets[key].source());
-        }
+        var hash = _loaderUtils2['default'].getHashDigest(compilation.assets[key].source(), 'sha1');
 
         _this4.hashesMap[hash] = validatedPath;
       });
-    }
-  }, {
-    key: 'findAssetsHashes',
-    value: function findAssetsHashes(compilation) {
-      var map = {};
-
-      compilation.chunks.forEach(function (chunk) {
-        if (chunk.hash && chunk.files.length) {
-          map[chunk.hash] = chunk.files[0];
-        }
-      });
-
-      return map;
     }
   }, {
     key: 'setNetworkOptions',
