+++ tests/legacy/fixtures/basic-with-runtime/__expected/appcache/manifest.appcache
@@ -1,5 +1,5 @@
 CACHE MANIFEST
-#ver:422fb99590c6ebbb3c26
+#ver:da39a3ee5e6b4b0d3255bfef95601890afd80709
 
 CACHE:
 ../external.js
