+++ tests/legacy/fixtures/basic-with-sw-custom-entry/__expected/appcache/manifest.html
