+++ tests/legacy/fixtures/basic-with-sw-custom-entry/__expected/main.js
