+++ tests/legacy/fixtures/basic-with-sw-custom-entry/sw-entry.js
