+++ tests/legacy/fixtures/basic-with-sw-custom-entry/webpack.config.js
