+++ tests/legacy/fixtures/cache-externals-absolute/__expected/sw.js
@@ -11,11 +11,11 @@ var __wpo = {
     "https://fonts.googleapis.com/css?family=Montserrat:400,700"
   ],
   "hashesMap": {
-    "c7afaa1605360b6eecceecbb43d62f68": "./main.js"
+    "fe8df8db17e5f8edb5c44ada1a7fdd9fe1a5012c": "./main.js"
   },
   "strategy": "changed",
   "responseStrategy": "cache-first",
-  "version": "3b59d46a1a71fdd7f971",
+  "version": "f2435b5e97ab165cde456b61eb28bf8bc103076e",
   "name": "webpack-offline",
   "relativePaths": true
 };
