+++ tests/legacy/fixtures/loaders-externals/__expected/appcache/manifest.html
