+++ tests/legacy/fixtures/loaders-externals/__expected/main.js
