+++ tests/legacy/fixtures/loaders-externals/webpack.config.js
