+++ tests/legacy/fixtures/paths-explicit-absolute-tools-override-outside/__expected/appcache/manifest.appcache
@@ -1,5 +1,5 @@
 CACHE MANIFEST
-#ver:ce2d8775efaeab7ec659
+#ver:9b833143ede63230057216fc5d687fdaa7067c93
 
 CACHE:
 /dist/main.js
