+++ tests/legacy/fixtures/paths-explicit-relative-inside/__expected/offline/sw.js
@@ -15,11 +15,11 @@ var __wpo = {
     "/images/logo_grey.svg"
   ],
   "hashesMap": {
-    "c7afaa1605360b6eecceecbb43d62f68": "../main.js"
+    "fe8df8db17e5f8edb5c44ada1a7fdd9fe1a5012c": "../main.js"
   },
   "strategy": "changed",
   "responseStrategy": "cache-first",
-  "version": "3b59d46a1a71fdd7f971",
+  "version": "f2435b5e97ab165cde456b61eb28bf8bc103076e",
   "name": "webpack-offline",
   "relativePaths": true
 };
