+++ tests/legacy/fixtures/paths-explicit-relative-override-outside/__expected/appcache/manifest.appcache
@@ -1,5 +1,5 @@
 CACHE MANIFEST
-#ver:9ac5b95d1b3b0ab68e3e
+#ver:ce7b708472540d89cb6e164003a65bd8772715dd
 
 CACHE:
 ../dist/main.js
