+++ tests/legacy/fixtures/paths-implicit-relative-inside/__expected/offline/appcache/manifest.appcache
@@ -1,5 +1,5 @@
 CACHE MANIFEST
-#ver:3b59d46a1a71fdd7f971
+#ver:f2435b5e97ab165cde456b61eb28bf8bc103076e
 
 CACHE:
 ../../main.js
