+++ tests/legacy/fixtures/update-strategy-all/__expected/appcache/manifest.appcache
@@ -1,5 +1,5 @@
 CACHE MANIFEST
-#ver:3b59d46a1a71fdd7f971
+#ver:da39a3ee5e6b4b0d3255bfef95601890afd80709
 
 CACHE:
 ../external.js
